refactor: unify tunnel models
