+++ linkup-clients/src/lib.rs
@@ -2,4 +2,4 @@ mod local_server;
 mod worker;
 
 pub use local_server::{Error as LocalServerClientError, LocalServerClient};
-pub use worker::{Error as WorkerClientError, TunnelData, WorkerClient};
+pub use worker::{Error as WorkerClientError, WorkerClient};
